% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{acipair}
\alias{acipair}
\title{Acitelli Data}
\format{
An object of class \code{"cross"}; see \code{\link[qtl]{read.cross}}.
}
\usage{
data(acipair)
}
\description{
Data from 148 heterosexual married couples in Detroit
}
\examples{
str(acipair)
}
\keyword{datasets}
