% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apim.R
\name{apim}
\alias{apim}
\title{actor-partner interdependence model

will run the apim on indistinguishable dyads eventualy 

even more stuff about the apim}
\usage{
apim(x, dat, dyadid, option = "indistinguishable")
}
\arguments{
\item{x}{is a function with form y ~ x.}

\item{dat}{is a dataset.}

\item{dyadid}{is a dyad in a dataset.}

\item{option}{is to see whether variables are indistinguishable}
}
\description{
actor-partner interdependence model

will run the apim on indistinguishable dyads eventualy 

even more stuff about the apim
}
\details{
some additional details about these functions
}
