% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confintLME.R
\name{confintLME}
\alias{confintLME}
\title{Confidence Intervals for LME Fixed Effects}
\usage{
confintLME(outp)
}
\arguments{
\item{outp}{the lme model object.}
}
\description{
Confidence Intervals for LME Fixed Effects
}
\details{
This function returns the confidence intervals in an organized fashion.
}
