% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_labels.R
\name{counts_labels}
\alias{counts_labels}
\title{counts_labels}
\usage{
counts_labels(data, x)
}
\arguments{
\item{data}{a data.frame}

\item{x}{a column in 'data'. Must be a string.}
}
\description{
counts_labels
}
