% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crsp.R
\name{crsp}
\alias{crsp}
\title{Multiple Pseudo R-squared

Get multiple R Squared}
\usage{
crsp(esdmod, esdemod)
}
\arguments{
\item{esdmod}{is error standard devisiation of the model.}
}
\details{
some additional details about these functions
}
