% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getevd.R
\name{getevd}
\alias{getevd}
\title{Error standard deviation for distinguishable dyads

Get error standard deviations from the distinguishable model with gls}
\usage{
getevd(outp, tipe = "gls", grp1 = "Group 1", grp2 = "Group 2")
}
\arguments{
\item{outp}{is the model object.  Not it is not the "summary" of that object.}

\item{tipe}{is the type of model object.  It defaults to "gls."  The other option is "lme."}

\item{grp1}{is the group 1 of the model object.(?)}

\item{grp2}{is the group 2 of the model object.(?)}
}
\description{
Error standard deviation for distinguishable dyads

Get error standard deviations from the distinguishable model with gls
}
\details{
some additional details about these functions
}
