% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lincomb.R
\name{lincomb}
\alias{lincomb}
\title{Tests of contrasts

Test the sum (S), the average (A), or the difference (D) of two effects from the same model.}
\usage{
lincomb(outp, v1, v2, fun = "D")
}
\arguments{
\item{outp}{is the summary of the model object. For exmaple, summary(mod). It can be a gls or lme object.}

\item{v1}{is the number of the first effect.}

\item{v2}{is the number of the second effect.}

\item{fun}{is the comparison. Default fun = "D". Other options include "S" the sum, and "A" the average.}
}
\details{
some additional details about these functions
}
