% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LRtestvar.R
\name{LRtestvar}
\alias{LRtestvar}
\alias{ranTestLR}
\title{Test the variance of a random effect}
\usage{
ranTestLR(outp, p, ctrl = defaultctrl)
}
\arguments{
\item{outp}{the lme model object.}

\item{p}{string variable that you want to test the random variance component for.}

\item{ctrl}{an lmeControl object. Defaults to msMaxIter=10000, MaxIter=100000, msMaxEval=10000, returnObject=TRUE, niterEM=10000, nlmStepMax=1000.}
}
\description{
Test the variance of a random effect
}
\details{
Example call: LRtestvar(dyadGC_full, "MTime").
}
