% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smallsummary.R
\name{smallsummary}
\alias{smallsummary}
\title{Gives a Smaller Summary of gls and lme Objects}
\usage{
smallsummary(outp)
}
\arguments{
\item{outp}{the lme or gls model object.}
}
\details{
This function returns the random effects (lme only), Rho, error variances, fixed effects, and their confidence intervals.
}
