% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sobel.R
\name{sobel}
\alias{sobel}
\title{Soboel Method for Testing Idirect Effect

The Sobel method is conservative. Use mmc, or bootstrapping if marginal. The Sobel test also presumes a and b are uncorrelated.}
\usage{
sobel(a, b, ase, bse)
}
\arguments{
\item{a}{path from X to M.}

\item{b}{path from M to Y.}

\item{ase}{standard error for path a.}

\item{bse}{standard error for path b.}
}
\details{
Could and maybe should output p value -- code available from last year
}
