% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testvar.R
\name{testvar}
\alias{testvar}
\title{Test the variance of a random effect}
\usage{
testvar(ovar, xx, yy, dset, DID, obs, dvv, ctrl = defaultctrl)
}
\arguments{
\item{ovar}{the outcome variable in quotes.}

\item{xx}{string set of fixed variables separated by comma (program assumes two intercept).}

\item{yy}{string set of random variables separated by comma (program assumes no  intercept in the usual sense); last item on the list is tested (the first two arguments are the same here, but they do not have to be; there might be more items in the fixed part.}

\item{dset}{the dataset.}

\item{DID}{the dyad ID in quotes.}

\item{obs}{the observation ID variable in quotes. This is a unique number for each dyad-measurement.}

\item{dvv}{the distinguishing variable in quotes.}

\item{ctrl}{an lmeControl object. Defaults to msMaxIter=10000, MaxIter=100000, msMaxEval=10000, returnObject=TRUE, niterEM=10000, nlmStepMax=1000.}
}
\description{
Test the variance of a random effect
}
\examples{
\dontrun{
testvar(
  "AConflict", "man,woman,MTime,WTime",
  "man,woman,WTime,MTime", "kashy_ppp", "DYADID", "obsid", "GenderS"
)
}
}
