% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_labels.R
\name{var_labels}
\alias{var_labels}
\title{var_labels}
\usage{
var_labels(data)
}
\arguments{
\item{data}{a data.frame}
}
\description{
Creates a `tbl_df` that contains the column names and variable labels to
easily explore the data.
}
