% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_view.R
\name{variable_view}
\alias{variable_view}
\title{Variable names with labels like in SPSS}
\usage{
variable_view(data)
}
\arguments{
\item{data}{the lme model object.}
}
\description{
Variable names with labels like in SPSS
}
\details{
Example call: variable_view(acipair)
}
