% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apim.R
\name{apim}
\alias{apim}
\title{actor-partner interdependence model

will run the apim on indistinguishable dyads eventualy

even more stuff about the apim}
\usage{
apim(x, dat, dyadid, option = "indistinguishable")
}
\arguments{
\item{x}{is a function with form \code{y ~ x}.}

\item{dat}{is the dataframe.}

\item{dyadid}{is a dyadic variable.}

\item{option}{is to check whether dyad is indistingusihable.}
}
\description{
actor-partner interdependence model

will run the apim on indistinguishable dyads eventualy

even more stuff about the apim
}
\details{
some additional details about these functions
}
