% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apim.R
\name{apim}
\alias{apim}
\title{actor-partner interdependence model

will run the apim on indistinguishable dyads

even more stuff about the apim}
\usage{
apim(x, dat, dyadid, option = "indistinguishable")
}
\arguments{
\item{x}{is a function}
}
\details{
some additional details about these functions
}

