% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counts_labels.R
\name{counts_labels}
\alias{counts_labels}
\title{counts_labels}
\usage{
counts_labels(data, x)
}
\arguments{
\item{data}{a data.frame}

\item{x}{a column in 'data'. Must be a string.}
}
\description{
Counts the number of observations in each category of a variable that has
value labels, returns counts and percent of total.
}
