% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crsp.R
\name{crsp}
\alias{crsp}
\title{Multiple Pseudo R-squared

Get multiple R Squared}
\usage{
crsp(esdmod, esdemod, indistinguishable = TRUE)
}
\arguments{
\item{esdmod}{is error standard deviation of the model of interest.}

\item{esdemod}{is error standard deviation of the empty model.}

\item{indistinguishable}{is to check whether the dyad is indistinguishable or not.}
}
\description{
Multiple Pseudo R-squared

Get multiple R Squared
}
\details{
some additional details about these functions
}
