% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getevd.R
\name{getevd}
\alias{getevd}
\title{Error standard deviation for distinguishable dyads

Get error standard deviations from the distinguishable model}
\usage{
getevd(outp)
}
\arguments{
\item{outp}{is the model object.}
}
\details{
some additional details about these functions
}
