% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmc.R
\name{mmc}
\alias{mmc}
\title{Monte Carlo confidence interval for the indirect effect}
\usage{
mmc(a, b, ase, bse, cov = 0, rep = 40000, conf = 95, dig = 6)
}
\arguments{
\item{a}{path from X to M.}

\item{b}{path from M to Y.}

\item{ase}{standard error for path a.}

\item{bse}{standard error for path b.}

\item{cov}{covariance for a and b (set to zero in MLM, but could be free if using lavaan). Defaults to zero.}

\item{rep}{number of replications (defaults to 40,000).}

\item{conf}{confidence level (defaults to 95).}

\item{dig}{number of decimal places (defaults to 6)}
}
\details{
Could and maybe should output p value -- code available from last year. Base code taken from Selig and Preacher (2008).
}
