% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sobel.R
\name{sobel}
\alias{sobel}
\title{Soboel Method for Testing Idirect Effect

The Sobel method is concervative. Use mmc, or bootstrapping if marginal.}
\usage{
sobel(a, b, asesq, bsesq)
}
\arguments{
\item{a}{path from X to M.}

\item{b}{path from M to Y.}

\item{asesq}{standard error for path a squared.}

\item{bsesq}{standard error for path b squared.}
}
\details{
Could and maybe should output p value -- code available from last year
}
