% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sobel.R
\name{sobel}
\alias{sobel}
\title{Sobel Method for Testing Idirect Effects}
\usage{
sobel(a, b, ase, bse)
}
\arguments{
\item{a}{path from X to M.}

\item{b}{path from M to Y.}

\item{ase}{standard error for path a.}

\item{bse}{standard error for path b.}
}
\description{
Sobel Method for Testing Idirect Effects
}
\details{
The Sobel method is conservative. Use the mmc function, or bootstrapping if marginal. The Sobel test also presumes a and b are uncorrelated.
}
